+++ src/component/navbars.jsx
@@ -9,7 +9,7 @@ import { Cooyub } from './icons';
 import ArConnect from './arconnect';
 import { Searchbar } from './search';
 import { appContext } from '../utils/initStateGen';
-import { LANGUAGES } from '../utils/ui';
+import { LANGUAGES } from '../utils/languages';
 
 export function Sidenav() {
   const { t, i18n } = useTranslation();
@@ -35,7 +35,7 @@ export function Sidenav() {
         <button className="w-9 h-9 btn btn-ghost btn-sm btn-square hover:text-zinc-200" onClick={() => switchView("featured")} style={{color: cond("/featured") ? 'white': ''}} disabled={cond("/featured") ? true: false}>
           <HomeIcon />
         </button>
-        <button className="w-9 h-9 btn btn-ghost btn-sm btn-square hover:text-zinc-200" onClick={() => switchView("following")} style={{color: cond("/following") ? 'white': ''}} disabled={cond("/following") ? true: false}>
+        <button className="w-9 h-9 btn btn-ghost btn-sm btn-square hover:text-zinc-200" onClick={() => switchView("following")} style={{color: cond("/following") ? 'white': ''}} disabled={cond("/following") ? true: true}>
           <CollectionIcon />
         </button>
         <div className="dropdown dropdown-hover mb-[-6px]">
